[#11] Refactor: 공통 모듈들 src/shared 디렉토리로 이동
