+++ src/shared/apis/apiClient.ts
@@ -1,4 +1,4 @@
-import { tokenStorage } from "@/lib/tokenStorage/tokenStorage";
+import { tokenStorage } from "@/shared/lib/tokenStorage/tokenStorage";
 import axios from "axios";
 
 const apiClient = axios.create({
@@ -15,7 +15,7 @@ apiClient.interceptors.request.use(
 
     return config;
   },
-  (error) => Promise.reject(error)
+  (error) => Promise.reject(error),
 );
 
 apiClient.interceptors.response.use(
@@ -62,7 +62,7 @@ apiClient.interceptors.response.use(
     }
 
     return Promise.reject(error);
-  }
+  },
 );
 
 export default apiClient;
