+++ src/shared/assets/images/android-icon-foreground.png
