+++ src/shared/assets/images/closed_fridge.png
