+++ src/shared/assets/images/icon.png
