+++ src/shared/assets/images/open_fridge.png
