+++ src/shared/assets/images/react-logo@3x.png
