+++ src/shared/components/AnimatedSplashScreen.tsx
