+++ src/shared/hooks/use-color-scheme.ts
@@ -1,0 +1,1 @@
+export { useColorScheme } from "react-native";
