+++ src/shared/scripts/reset-project.js
